source/requirements.txt
SQLAlchemy==1.3.19
PyMySQL==0.10.1
reportlab==3.5.48
cryptography==3.0
Unidecode==1.1.1
Pillow==7.2.0
